pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = ""
authors = [
    {name = "chaimajaballah0",email = "chaimajaballah57@gmail.com"}
]


readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "langchain (>=0.3.25,<0.4.0)",
    "langgraph (>=0.4.1,<0.5.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "google-genai (>=1.13.0,<2.0.0)",
    "mcp[cli] (>=1.7.1,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "langchain-google-community (>=2.0.7,<3.0.0)",
    "langchain-google-genai (>=2.1.4,<3.0.0)",
    "google-api-python-client (>=2.169.0,<3.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
    "typing-extensions (>=4.13.2,<5.0.0)",
    "langchain-mcp-adapters (>=0.0.11,<0.0.12)",
    "langgraph-prebuilt (>=0.1.7,<0.2.0)",
    "langgraph-supervisor (>=0.0.21,<0.0.22)",
    "langchain-google-vertexai (>=2.0.23,<3.0.0)",
    "google-cloud (>=0.34.0,<0.35.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "langchain-core (>=0.3.59,<0.4.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "uuid (>=1.30,<2.0)",
    "selenium (>=4.32.0,<5.0.0)",
    "nest-asyncio (>=1.6.0,<2.0.0)",
    "langsmith (>=0.3.42,<0.4.0)",
]
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[tool.poetry]
packages = [
    { include = "authentication", from = "src" },
    { include = "client", from = "src" },
    { include = "servers", from = "src" },
]